refactor(connector): do not expose internal implementation for parser benchmarks
